+++ ntex/tests/http_ws_client.rs
@@ -41,7 +41,7 @@ async fn test_simple() {
 
                     // start websocket service
                     Dispatcher::new(
-                        io.into_boxed(),
+                        io.seal(),
                         ws::Codec::default(),
                         ws_service,
                         Default::default(),
@@ -53,7 +53,7 @@ async fn test_simple() {
     });
 
     // client service
-    let (_, io, codec) = srv.ws().await.unwrap().into_inner();
+    let (io, codec, _) = srv.ws().await.unwrap().into_inner();
     io.send(&codec, ws::Message::Text(ByteString::from_static("text")))
         .await
         .unwrap();